Manage permission account

+++ source_code/app/Http/Controllers/Admin/AdminFacultyController.php
@@ -9,6 +9,7 @@ use App\Company;
 use DataTables;
 use Validator;
 use App\Tag;
+use Auth;
 class AdminFacultyController extends Controller
 {
     /**
@@ -18,39 +19,36 @@ class AdminFacultyController extends Controller
      */
     public function index()
     {
-        // $faculties = Faculty::with('tags')->get();
-        // $faculties = Faculty::all();
-        // foreach ($faculties as $faculty) {
-        //     return $faculty;
-        // }
-        // return view('admin.faculties.index', compact('faculties'));
-        return view('admin.faculties.index');
+        if(Auth::user()->can('faculties.view')){
+            return view('admin.faculties.index');
+        }
+        else{
+            return view('errors.admin_auth');
+        }
     }
 
     function getdata()
     {
-     $faculties = Faculty::with('tags');
-     return DataTables()::of($faculties)->addColumn('action', function($faculty){
-        return '<a href="#" class="btn btn-xs btn-primary edit" id="'.$faculty->id.'"><i class="fa fa-pencil-square-o" aria-hidden="true"></i>Edit</a>
-        <a href="#" class="btn btn-xs btn-danger delete" id="'.$faculty->id.'"><i class="fa fa-trash" aria-hidden="true"></i> Delete</a>
-        ';
-    })
-        //  ->addColumn('tag', function($faculty)
-        //   {
-        //     $tmp = '';
-        //     foreach ($faculty->tags as $tag)
-        //     {
-        //         $tmp .=  '<span class="label label-default">'.$tag->name.'</span>';
-        //     }
-        //     return $tmp;
-        // })->rawColumns(['tag', 'action']) 
-     ->make(true);
- }
+        if(Auth::user()->can('faculties.view'))
+            {
+               $faculties = Faculty::with('tags');
+               return DataTables()::of($faculties)->addColumn('action', function($faculty){
+                $tmp = '';
+                if(Auth::user()->can('faculties.update')){
+                    $tmp .='<a href="#" class="btn btn-xs btn-primary edit" id="'.$faculty->id.'"><i class="fa fa-pencil-square-o" aria-hidden="true"></i>Edit</a>';
+                }
+                if(Auth::user()->can('faculties.delete')){
+                    $tmp .='<a href="#" class="btn btn-xs btn-danger delete" id="'.$faculty->id.'"><i class="fa fa-trash" aria-hidden="true"></i> Delete</a>';
+                }
+                // return '<a href="#" class="btn btn-xs btn-primary edit" id="'.$faculty->id.'"><i class="fa fa-pencil-square-o" aria-hidden="true"></i>Edit</a>
+                // <a href="#" class="btn btn-xs btn-danger delete" id="'.$faculty->id.'"><i class="fa fa-trash" aria-hidden="true"></i> Delete</a>
+                // ';
+                return $tmp;
+            })->make(true);
+           }else
+           {return response()->json('error', 400);}
 
- public function removedata(Request $request)
- {
-    return '123';
-}
+       }
 
     /**
      * Show the form for creating a new resource.
@@ -59,7 +57,10 @@ class AdminFacultyController extends Controller
      */
     public function create()
     {
-        return view('admin.faculties.create');
+        if(Auth::user()->can('faculties.create')){
+            return view('admin.faculties.create');
+        }
+        else{return view('errors.admin_auth');}
     }
 
     /**
@@ -70,36 +71,42 @@ class AdminFacultyController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,[
-            'name'=>'required|unique:faculties,name',
-            'description'=>'required',
-        ]);
-        $tags = explode(',', request('tags')); 
-        $data = [
-            'name'=>request('name'),
-            'description'=>request('description'),
-        ];
+        if(Auth::user()->can('faculties.create')){
+            $this->validate($request,[
+                'name'=>'required|unique:faculties,name',
+                'description'=>'required',
+            ]);
+            $tags = explode(',', request('tags')); 
+            $data = [
+                'name'=>request('name'),
+                'description'=>request('description'),
+            ];
 
-        $faculty = Faculty::create($data);
+            $faculty = Faculty::create($data);
 
-        /*Save tagss*/
-        if($request->tags)
-        {
-            foreach ($tags as $key => $value) 
+            /*Save tagss*/
+            if($request->tags)
             {
-                if(count(Tag::Where('name',$value)->get()) !=0)
-                    {
-                        $faculty->tags()->save(Tag::where('name',$value)->first());
-                    }
-                    else
-                    {
-                        $tg = Tag::create(['name'=>$value]);
-                        $faculty->tags()->save($tg);
+                foreach ($tags as $key => $value) 
+                {
+                    if(count(Tag::Where('name',$value)->get()) !=0)
+                        {
+                            $faculty->tags()->save(Tag::where('name',$value)->first());
+                        }
+                        else
+                        {
+                            $tg = Tag::create(['name'=>$value]);
+                            $faculty->tags()->save($tg);
+                        }
                     }
                 }
+                return redirect(route('faculties.index'))->with('message','Tạo mới Ngành nghề thành công!');
             }
-            return redirect(route('faculties.index'))->with('message','Tạo mới Ngành nghề thành công!');
-        }
+            else{
+             return view('errors.admin_auth');
+         }
+
+     }
 
     /**
      * Display the specified resource.
@@ -132,66 +139,70 @@ class AdminFacultyController extends Controller
      */
     public function update(Request $request)
     {
-        $validation = Validator::make($request->all(), [
-            'name' => 'required',
-            'description'  => 'required',
-        ]);
-        
-        $error_array = array();
-        $success_output = '';
-        if ($validation->fails())
-        {
-            foreach ($validation->messages()->getMessages() as $field_name => $messages)
-            {
-                $error_array[] = $messages; 
-            }
-        }
-        else
-        {
-            if($request->get('button_action') == 'insert')
-            {
+        if(Auth::user()->can('faculties.update')){
+            $validation = Validator::make($request->all(), [
+                'name' => 'required',
+                'description'  => 'required',
+            ]);
 
-            }
-
-            if($request->get('button_action') == 'update')
+            $error_array = array();
+            $success_output = '';
+            if ($validation->fails())
             {
-                $faculty = Faculty::find($request->get('faculty_id'));
-                $faculty->name = $request->get('name');
-                $faculty->description = $request->get('description');
-                $tags1 = explode(',', request('tags')); 
-                $tags2;
+                foreach ($validation->messages()->getMessages() as $field_name => $messages)
+                {
+                    $error_array[] = $messages; 
+                }
+            }
+            else
+            {
+                if($request->get('button_action') == 'insert')
+                {
+
+                }
+
+                if($request->get('button_action') == 'update')
+                {
+                    $faculty = Faculty::find($request->get('faculty_id'));
+                    $faculty->name = $request->get('name');
+                    $faculty->description = $request->get('description');
+                    $tags1 = explode(',', request('tags')); 
+                    $tags2;
                     //change right id key tags
-                foreach ($tags1 as $key => $value) {
-                    $tmpTag = Tag::where('name', $value)->get();
+                    foreach ($tags1 as $key => $value) {
+                        $tmpTag = Tag::where('name', $value)->get();
                         // return $tmpTag->first()['id'];
                         // $tmpTag = Tag::where('name', $value)->first(['id'])['id'];
-                    if(count($tmpTag) !=0)
-                    {
-                        $id =  $tmpTag->first()['id'];
-                        $tags2[$id] = $tags1[$key];
+                        if(count($tmpTag) !=0)
+                        {
+                            $id =  $tmpTag->first()['id'];
+                            $tags2[$id] = $tags1[$key];
+                        }
+                        else
+                        {
+                            $tg = Tag::create(['name'=>$value]);
+                            $tags2[$tg->id] = $tags1[$key];
+                        }
                     }
-                    else
-                    {
-                        $tg = Tag::create(['name'=>$value]);
-                        $tags2[$tg->id] = $tags1[$key];
-                    }
-                }
                     //update tags
-                $faculty->tags()->sync(array_keys($tags2));
-                $faculty->save();
-                $success_output = '<div class="alert alert-success">Data Updated</div>';
-            }
-            
-        }
-        
-        $output = array(
-            'error'     =>  $error_array,
-            'success'   =>  $success_output,
-            'faculty'   =>  $faculty,
-            'tag'       => $tags2
-        );
-        return response()->json($output);
+                    $faculty->tags()->sync(array_keys($tags2));
+                    $faculty->save();
+                    $success_output = '<div class="alert alert-success">Data Updated</div>';
+                }
 
+            }
+
+            $output = array(
+                'error'     =>  $error_array,
+                'success'   =>  $success_output,
+                'faculty'   =>  $faculty,
+                'tag'       => $tags2
+            );
+            return response()->json($output);
+        }
+        else{
+            return response()->json('error',400);
+        }
     }
 
     /**
@@ -202,13 +213,17 @@ class AdminFacultyController extends Controller
      */
     public function destroy(Request $request)
     {
-        if(Faculty::find($request->id)->delete())
-            {
-                return response()->json('success');
-            }
-            else
-            {
-               return response()->json('error');
-           }
-       }
-   }
+        if(Auth::user()->can('faculties.delete')){
+            if(Faculty::find($request->id)->delete())
+                {
+                    return response()->json('success');
+                }
+                else
+                {
+                 return response()->json('error',400);
+             }
+         }
+         else{return response()->json('error',400);}
+
+     }
+ }
